requirements.txt
pandas==0.25.0
paramiko>=2.4.2
tqdm==4.23.4
PyYAML==5.3.1